refactor #211 : 더보기 FAB 위치 계산을 섹션 기준으로 변경 및 분실물 목록 첫 카드 상단 테두리 수정
