+++ src/component/root/layout/Sidebar.tsx
@@ -1,6 +1,6 @@
 import Authentication from './Authentication';
 import Logo from './Logo';
-import { useEffect, useRef, useState } from 'react';
+import { useEffect } from 'react';
 import { useLocation, useNavigate } from 'react-router-dom';
 import {
   Map,
@@ -8,15 +8,11 @@ import {
   CircleUser,
   Ellipsis,
   X,
-  BookOpen,
-  Users,
-  MessageCircleMore,
   House,
 } from 'lucide-react';
 import {
   useSelectedMode,
   useSetSelectedMode,
-  useBottomSheetOpen,
 } from '../../../store/hooks/useMainStore';
 import { useAuthFlag } from '../../../store/hooks/useAuth';
 import { clearFormData } from '../../../utils/register/registerStorage';
@@ -27,25 +23,11 @@ import { SECTION_MODE_MAP } from '../../onboarding/onboardingSteps';
 const Sidebar = () => {
   const selectedMode = useSelectedMode();
   const setSelectedMode = useSetSelectedMode();
-  const isBottomSheetOpen = useBottomSheetOpen();
   const isAuthenticated = useAuthFlag();
   const navigate = useNavigate();
   const { pathname } = useLocation();
   const tourSectionId = useOnboardingStore((s) => s.tourSectionId);
   const endTour = useOnboardingStore((s) => s.actions.endTour);
-  const [isMoreOpen, setIsMoreOpen] = useState(false);
-  const moreFabRef = useRef<HTMLDivElement>(null);
-
-  useEffect(() => {
-    if (!isMoreOpen) return;
-    const handleClickOutside = (e: MouseEvent) => {
-      if (moreFabRef.current && !moreFabRef.current.contains(e.target as Node)) {
-        setIsMoreOpen(false);
-      }
-    };
-    document.addEventListener('mousedown', handleClickOutside);
-    return () => document.removeEventListener('mousedown', handleClickOutside);
-  }, [isMoreOpen]);
 
   useEffect(() => {
     if (pathname.startsWith('/register')) {
@@ -168,109 +150,6 @@ const Sidebar = () => {
         {/* 5) 로그인/로그아웃 */}
         <Authentication />
       </div>
-
-      {/* 모바일 전용: 더보기 FAB (위저드 플로우 진입 시 숨김) */}
-      {(selectedMode === 'find' || selectedMode === 'register') &&
-        !pathname.startsWith('/find') &&
-        !pathname.startsWith('/register') && (
-          <div
-            ref={moreFabRef}
-            className={`fixed right-4 z-50 flex gap-2 transition-[bottom] duration-300 ease-out md:hidden ${
-              isBottomSheetOpen ? 'flex-row items-center' : 'flex-col items-end'
-            }`}
-            style={{
-              bottom:
-                selectedMode === 'find'
-                  ? isBottomSheetOpen
-                    ? 'calc(72vh)'
-                    : 'calc(3.5rem + 64px + 1rem)'
-                  : '5rem',
-            }}
-          >
-            {[
-              {
-                label: '가이드',
-                icon: <BookOpen className="h-5 w-5" />,
-                onClick: () => {
-                  navigate('/onboarding');
-                  setIsMoreOpen(false);
-                },
-              },
-              {
-                label: '팀 소개',
-                icon: <Users className="h-5 w-5" />,
-                onClick: () => {
-                  navigate('/more/team');
-                  setIsMoreOpen(false);
-                },
-              },
-              {
-                label: '피드백',
-                icon: <MessageCircleMore className="h-5 w-5" />,
-                href: 'https://forms.gle/xzvHjcbKh3vumBXQA',
-              },
-            ].map((item, i, arr) => {
-              // 수평(시트 열림): 오른쪽(FAB)에서 가까운 항목부터 등장
-              const openDelay = isBottomSheetOpen
-                ? (arr.length - 1 - i) * 50
-                : (arr.length - 1 - i) * 50;
-              const closeDelay = isBottomSheetOpen ? i * 50 : i * 50;
-              const hiddenClass = isBottomSheetOpen
-                ? 'pointer-events-none translate-x-4 opacity-0'
-                : 'pointer-events-none translate-y-4 opacity-0';
-
-              return (
-                <div
-                  key={item.label}
-                  style={{ transitionDelay: `${isMoreOpen ? openDelay : closeDelay}ms` }}
-                  className={`transition-all duration-200 ease-out ${
-                    isMoreOpen ? 'translate-x-0 translate-y-0 opacity-100' : hiddenClass
-                  }`}
-                >
-                  {item.href ? (
-                    <a
-                      href={item.href}
-                      target="_blank"
-                      rel="noopener noreferrer"
-                      onClick={() => setIsMoreOpen(false)}
-                      className="flex flex-col items-center gap-1"
-                    >
-                      <div className="flex h-12 w-12 items-center justify-center rounded-full bg-white text-teal-500 shadow-md transition hover:bg-teal-50">
-                        {item.icon}
-                      </div>
-                      <span className="text-xs text-gray-600">{item.label}</span>
-                    </a>
-                  ) : (
-                    <button onClick={item.onClick} className="flex flex-col items-center gap-1">
-                      <div className="flex h-12 w-12 items-center justify-center rounded-full bg-white text-teal-500 shadow-md transition hover:bg-teal-50">
-                        {item.icon}
-                      </div>
-                      <span className="text-xs text-gray-600">{item.label}</span>
-                    </button>
-                  )}
-                </div>
-              );
-            })}
-            <button
-              data-tour="mobile-sidebar-more"
-              onClick={() => setIsMoreOpen((v) => !v)}
-              className={`flex h-12 w-12 shrink-0 items-center justify-center rounded-full border shadow-lg transition duration-200 active:scale-95 ${
-                isMoreOpen
-                  ? 'border-teal-500 bg-teal-500 text-white'
-                  : activeMore
-                    ? 'border-gray-300 bg-white text-teal-700'
-                    : 'border-gray-300 bg-white text-gray-600 hover:text-teal-500'
-              }`}
-              aria-label="더보기"
-            >
-              {isMoreOpen ? (
-                <X className="h-5 w-5 transition-transform duration-200" />
-              ) : (
-                <Ellipsis className="h-5 w-5 transition-transform duration-200" />
-              )}
-            </button>
-          </div>
-        )}
 
       {/* ---------- 데스크탑(md+): 세로 레이아웃 ---------- */}
       <div className="hidden h-full flex-col md:flex">
